backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai-whisper==20231117
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
python-socketio==5.10.0
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.27.2
openai==1.35.10
groq==0.4.1
